Se realiza cambios y agregó, mi parte del proyectó #146
Se agrego la parte 5, Creación de la Clase Conexion: Parte 3. También se elimino archivos, estos fuero: semana9.py y semana9 . Se pasaron archivos sueltos que estaban en la carpeta JavaScript a otra carpeta, dentro de la misma.
