+++ JavaScript/Objetos/Captura del codigo javascript.PNG
