+++ PYTHON/Capa de Datos Logging y Postgresql Parte 2/Persona.py
@@ -1,5 +1,7 @@
+from logger_base import log
+
 class Persona:
-    def __init__(self, id_persona, nombre, apellido, email):
+    def __init__(self, id_persona=None, nombre=None, apellido=None, email=None):
         self._id_persona = id_persona
         self._nombre = nombre
         self._apellido = apellido
@@ -45,3 +47,11 @@ class Persona:
     @email.setter
     def email(self, email):
         self._email = email
+
+if __name__ == '__main__':
+    persona1 = Persona(1, 'Juan', 'Perez', 'jperez@gmail.com')
+    log.debug(persona1)
+    persona2 = Persona(nombre='Jose', apellido='Lepez', email='ljose@gmail.com')
+    log.debug(persona2)
+    persona1 = Persona(id_persona = 1)
+    log.debug(persona1)
