+++ PYTHON/Capa de Datos Logging y Postgresql Parte 2/conexion.py
@@ -30,3 +30,20 @@ class Conexion:
             else:
                 return cls._conexion
 
+    @classmethod
+    def obtenerCursor(cls):
+        if cls._cursor is None:
+            try:
+                cls._cursor = cls.obtenerConexion().cursor()
+                log.debug(f'Se abrió corectamente el cursor: {cls._cursor}')
+                return cls._cursor
+            except Exception as e:
+                log.error(f'Ocurrio un error: {e}')
+                sys.exit()
+            else:
+                return cls._cursor
+
+
+if __name__ == '__main__':
+    Conexion.obtenerConexion()
+    Conexion.obtenerCursor()
